More robust handling of cities list in Location Dialog.  The previous implementation was pretty kludgy: we kept an array of integers that represented the index numbers in the master city list of each city in the (possibly filtered) list box of the dialog.  We then tried to match that integer back to the master list to identify cities in the list box. This worked, until we re-sorted the listbox contents without re-ordering the array of integer IDs.  oops.
Instead, we now keep a QPtrList in LocationDialog for the "filtered"
cities.  Items in the list box are just name strings, so to match a name
in the listbox to a city in the filtered list, we still must match by
name.  So, I added a function GeoLocation::fullName() which returns the
"City, Province, Country" string which is used to construct the listbox
contents and match the listbox items to actual cities.

The bottom line is, it actually works now, and the code is much less
ugly too.

Thanks for the bug report (I fixed the Massachusetts typo too).

CCMAIL: kstars-devel@kde.org
BUG: 97819

svn path=/trunk/kdeedu/kstars/; revision=382197