Fixed a bug that I should have caught long ago...
When computing the rise/set times for an object, we use a function
SkyObject::recomputeCoords() to temporarily reposition the object.  This
function restored the original coordinates of the object with "set(
originalRA, originalDec )".

However, the "set()" function sets the *J2000* coordinates, whereas
originalRA and originalDec were the precessed coords for the present
epoch!  So, any time rise/set times were computed, the J2000 coordinates
were reset to a wrong value!

I noticed this when trying to retrieve images from the Digitized Sky
Survey; the returned images were slightly off-center.

The fix was to use "setRA( originalRA )" and "setDec( originalDec )"
instead.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=382528