Finally, the fix we've all been waiting for! :)
When the edge of the visible sky circle is visible, the opaque ground polygon
always looked strange, because it was extended beyond the bounds of the sky
circle to the edge of the window.  This commit makes the ground polygon
follow the bottom edge of the sky circle instead.

But wait, there's more!  In addition, if you cvs up now, you'll get
another fix *absolutely free*!  When the opaque ground is drawn, objects
below the horizon are now immediately rejected by checkVisibility(), which is
a nice reduction in CPU usage.  We had this check in before, but I commented
it out because if you zoom in on a point below the horizon, the ground
polygon disappears.  So I added some fancier logic to checkVisibility() to
determine if the ground is visible or not.

cvs up now, servers are standing by...

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=391450