Fix crash when opening Details window on solar system body (undefined pointer).
Some cosmetic fixes to the Details window: move object name to titlebar; move
image to bottom of widget.

make titlebars of links and log tabs black (actually, "Text" color).  Maybe
we should do this for all of the boxes?  Or use the window titlebar color?

Add slots for the "Add to List", "Center" and "Scope" buttons in first tab.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=394005