Added Google image search to the Thumbnail image selector tool.  Pretty neat!
It isn't perfect; try opening the details window for IC 440, and click on the
image box to load the image search.  None of the images found are actually of
the object.  Any ideas how to improve this?  It currently searches for '"long
name" "primary name"' if both exist, and "primary name" if there is no long
name.

Do you think we need to inform/warn users that an image search is taking
place here?  Is it important to provide an option to [dis|en]able image
search in this tool?

TODO: if an image isn't found, I'd like to suppress the message box.
TODO: if the image is smaller than 200x200, need to fill the background.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=396014