VLSR calculator module:
  Added an implementation of the Ron & Vondrak method to compute the velocity
of the Earth referred to the barycenter of the solar system in the J2000
equatorial system. The velocity is given in X, Y and Z components an the units are
km/s. These values may be used for computing the radial velocity of an astronomical source
referred to the center of the Earth. It might also be used to compute the anual aberration,
although it is not being used now for that aim.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=399209