VLSR calculator module
Added a function TopocentricVelocity() which returns, in the parameter list, an array with thevelocity of a point on the surface of the Earth in km/s, for a given date and time. This value may
be used for computing the radial velocity of a source referred to a local point on the surface
of the Earth once the LSR velocity of that source is known. The X and Y axis are contained in
the equator and the X axis is towards the nodes line. The Z axis is along the poles.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=399211