VLSR calculator module
Added a calculator module which computes the radial velocity of an astronomical source
referred to the barycenter of the solar system, the center of the Earth and a local site
on the surface of the Earth once the LSR velocity of the source is known. In order to make these
computations I have added new functions in classes KSNumbers, SkypPoint and GeoLocation. The UI for
this module contains an additional tab which I have named "Overview". This tab contains a KTextRich
widget which explains what the modules does and how it works. I am trying to use PNG images to
display equations and figures. My intentionn is to extend this tab to most of the calculator
modules to help the reader understand what the module does.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=399220