Changes to fix some issues regarding the computation of the radial velocity of an astronomical source referred to the solar system barycenter, the center of the Earth and the observer's site from the radial velocity referred to the LSR.
I have included precession in the the SkyPoint class methods since the solar apex is in J2000
and we need to precess the source coordinates to J2000.

The geocentric velocity also needs the source coordinates to be in J2000 since the velocity
of the earth is a vector in the J2000 equinox coordinate system and therefore I have
also included precession in that function.

I have changed the parameters in the vHeliocentric and vGeocentric functions. Instead of
using the epoch I use the julian day. That reduces three parameters in the vGeocentric function to
two. In this latter function I also pass the vhelio as parameter. That means that one needs to compute
the heliocentric velocity before calling this method, and therefore we do not need to call this method
from within vGeocentric(). In the calculator this means that the vHeliocentric() method is called once
and not twice as before.

I have also added another method vTopocentric() which computes the topocentric velocity for
a given time once we know the geocentric velocity.

The module takes into account all these changes in the API of the skypoint class.

I have also rewritten the overview text but it is not ready yet. I still need to polish the text
and include the formulae inside the KRichText.
CCMAIL:kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=404586