Adding ability to load/save observing lists.  Please test if you have time.  I'd like opinions on some design decisions in this feature:
+ When saving an observing list, should we ask the user to specify the
filename (as we do for DCOP scripts), or simply create a filename from
the user's list name, and place it in $KDEHOME/share/apps/kstars (as we
do for color schemes and other stuff)?  Right now, I am asking for the
filename, but I think it might be better to just "hide" the actual files
in $KDEHOME.

+ When opening an observing list, if there are already objects in the
current list, should the new observing list be appended to the current
one, or should the current list be cleared first (with save warning, if
applicable)?  Right now, I have it clear the current list before loading
a new one.

+ Any brilliant ideas to improve the UI?  I'm not very excited about
having two rows of buttons at the top of the window.

thanks for any input!

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=408889