Adding new Custom catalogs tool, accessible from the Catalogs tab of the Configure window.  To add a custom catalog, first construct a file containing the catalog data.  The file should contain space-separated column data for each object.  Any of the following data fields can be used, in any order:
+ ID number
+ RA
+ Dec
+ Object type identifier
+ magnitude
+ major axis (arcmin)
+ minor axis (arcmin)
+ position angle (degrees)
+ common name / alternate name (wrap in quotes if it contains spaces!)
+ Ignore (a field in the data file that is to be skipped)

Open the Add Custom Catalog tool, and fill in the name of the catalog, 
the output catalog name, and the input data file you just created.  You 
need to arrange the "Catalog Fields" list, so that it contains the same 
data fields as your data file, and in the same order.  Just drag and 
drop.  For all fields except ignore, only one copy is allowed between 
the "Catalog fields" and "Available fields" boxes.  

Finally, select a coordinate epoch, a symbol color, and the catalog 
prefix.  The prefix is used with the ID numbers (if that is one of your 
fields) to construct the object name.  So if you use ID numbers, and 
your prefix is "YY", then the objects are named "YY 1", "YY 2", etc.

You can press the Preview button to make sure everything's okay.  If 
there are problems, you'll get a window explaining in some detail what 
the issues are.  When satisfied, press Ok, and your new catalog 
appears in the list.  Your objects are in the sky, and usable 
throughout the program.  The output catalog file is the same as the 
input data file, with a "metadata header" prepended to it that allows 
KStars to parse the file.

It's a pretty big commit, so please try it out and let me know what you 
think and if there are issues.  To get you started, I will post a 
sample catalog to the mailing list that you can play with.

KNOWN ISSUES: 

The difference between "Load Catalog" and "Create Catalog" may not be 
very obvious.  "Load Catalog" is only to be used if you have a file 
that already has a valid metadata header.

Tab order:  for some reason, the "Append existing file" edit box never 
gets input focus for tabbing.  The little "open file" icon next to it 
gets focus instead.

Need to add a mechanism to require RA and Dec fields in the 
catalog.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=422195