Ladies and Gentlemen, I give you:  Pluto the asteroid.
Yes, it's true, Pluto is now derived from KSAsteroid instead of
KSPlanetBase.  Using the asteroid-like orbital elements results in a
much more precise determination of its position, which remains robust
over a much longer interval of time (thousands of years instead of
~100 years).

The derived class (KSPluto) hard-codes the orbital elements for Pluto,
as taken from this site: http://ssd.jpl.nasa.gov/elem_planets.html It
also allows for the orbital elements to evolve with time, at a
constant rate.  The rates are also from that website, and are also
hard-coded in KSPluto,

Updated Pluto's orbital track data for the solar system viewer.  Also
fixed the calculation of the planets' X/Y coordinates in the tool.
Before, we gave X and Y in the plane of the planet's orbit, rather
than in the ecliptic plane.  The positions are now correctly projected
onto the ecliptic plane.  It only makes a noticeable difference for
Pluto (due to its large inclination angle).

Also fixed calculation of planet magnitudes.  I was getting "nan" mags
for both Pluto and Mercury.  I didn't understant the calculation of
the planet phase, so I used a much simpler formulation of it from
"Practical Astronomy with your Calculator".  Pablo, please have a look
at what I did there (in KSPlanetBase::findMagnitude).

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=423925