Fixing bug #108480: negative timesteps in the solar system viewer
Added a "bool DaysOnly" member to the timestep widget.  When true, all 
timesteps smaller than 1 day are removed from the list of possible 
timesteps.  This was necessary in order to make negative timesteps 
available in the solar system tool without having to show all the steps 
smaller than a day, which are useless in this context.

The TimeStepBox ctor now has an optional argument to determine if 
it is "days only" or not, and the days-only setting can be changed at 
any time after construction with TimeStepBox::setDaysOnly(bool).

BUG: 108480
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=432407