Possible fix for bug #106501 (crash while downloading Steinicke NGC catalog from GHNS)
I never confirmed a crash, but I did see odd behavior in downloading the 
Steinicke catalog:  console output indicates that it reads the new 
catalog into memory three times, which causes a long UI delay.  I tried 
to understand why it's happening, but I don't think I succeeded.  As a 
workaround, I added a bool member "NGCUpdated" to KSNewStuff, which 
starts out false and is set true by the installation of the Steinicke 
NGC catalog.  The installation will not be attempted if NGCUpdated is 
true.

Please test to see if this resolves your crash condition.

BUG: 106501
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=432669