I fixed the problem of telescope crosshair centering and tracking. I simply needed to call telescopeSkyObject->EquatorialToHorizontal() to update the telescope skyobject internal coords.
I also added Argo Navis to the list of supported telescopes, since the
controller was confirmed to work under KStars using the LX200 Basic
driver as report by Douglas Phillipson.

The context-menu entry for centering and tracking the telescope
crosshair was initially called "Center & Track" but I thought that might
confuse the user, he/she would think this action will somehow center and
track the telescope itself, and not simply the KStars crosshair. So I changed
it to "Center & Track Crosshair". It is more verbose, maybe too verbose
for a menu entry? But it certainly removes the ambiguity.

CCMAIL: kstars-devel@kde.org
CCMAIL: dougp@intermind.net

svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=441763