Fixing Bug #109788 (Rounding error when computing arcmin/arcsec for a dms angle)
The issue manifests itself when converting a double value with all 
zeroes after the decimal point, to an int.  e.g.:

  double d = 10.000000000;
  int i = int(d);  // i == 9 !! >:(

I tried several ways of massaging the code, but in the end, this is 
the only thing I could get to work:

  double d = 10.000000000;
  int i = int( float(d) );  // i == 10 !! :D
  
So, it works, but maybe it's a kludge and maybe someone knows a better 
way.

Will port to trunk.

CCMAIL: kstars-devel@kde.org
BUG: 109788


svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=442973