This is the initial import of breaking KStarsData and SkyMap into manageable pieces by building a self organized object hierarchy. Currently there are the basic classes SkyComponent and SkyComposite. The code for following classes is ported too: horizon, milky way, coordinate grid, constellation lines and constellation names.
The code will not compile, because it's mainly copy-and-pasted from KStarsData/SkyMap. The focus of this import is a to have a stable API. After having all code ported to the new infrastructure I will start to polish the code, so that it will compile. Ported code is just commented out in KStarsData/SkyMap and not entirely removed.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=445142