OK, here comes next step toward the new component structure. This time I've ripped off the code of deep sky objects and custom catalogs. It's a lot of stuff in kstarsdata right now, ~800 lines of code.
Now there are only left the planets + sun + moon, the asteroids and the comets which could share a lot of code. This is next on my list to solve this sanely. Oh, and equator and ecliptic are missing too, but these ones are not that hard to solve.

And as allways, nothing would compile, because it needs to add includes and so on, but hopefully this will happening soon, when libkdeedu will compile again and we can start the port to QT4 :-)

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=453270