Fix some refraction-related inconsistincies:
Display the "apparent" altitude (i.e., including refraction) in the
statusbar and focus infobox.

The displayed RA/Dec coords in the infobox were not correct when
refraction was active.

When switching between Equatorial and Horizontal coordinates, the
focus position would be slightly offset because it was not properly
accounting for refraction.

TODO: the refract(dms*, bool) function determines the apparent altitude
from the true altitude when invoked as refract(dms*, true), and
determines the true altitude from apparent altitude when invoked as
refract(dms, false).  However, there seems to be a slight
inconsistency:

  dms a( 45.0 );
  dms b = refract( &a, true );
  dms c = refract( &b, false );
  //c should be equal to a, but it is slightly smaller

CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=457604