Probable fix for bug #112925 (KStars crashes on startup).  Thanks to the detailed bug report, this was pretty easy to track down.  We use the untranslated object names internally, so when objects are created, the name parameter is supposed to be passed in English.  For user-visible uses, we are supposed to use SkyObject::translatedName().
The crash was caused by my use of i18n("Pluto") for the name parameter 
in the ctor of the new Pluto class, which sets the internal Name 
parameter to the localized name for Pluto.  Then when 
SkyMap::drawSolarSystem() calls PCat->findByName("Pluto"), it gets NULL 
instead of a pointer to Pluto, which causes a crash.  

I've replaced i18n("Pluto") with I18n_NOOP("Pluto") in the KSPluto ctor, 
which seems to work for the other planets.  Note that this fix may 
cause a fuzzy string.  Please apply the patch and test it.  If it works, 
I'll close the bug.  

CCMAIL: kstars-devel@kde.org
CCBUG: 112925

svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=463106