Fixing remaining issue in BR #112925.  The reported issue ("Moon not found" messages at startup) was actually just a symptom of the true problem: it wasn't just the Moon, if using a non-English language, then any object whose localized name starts with a different letter than its English name would be impossible to locate in the ObjectNameList.
This is due to an inconsistency in that class.  The class divides all 
named objects into 26 lists, to make locating an object by its name 
much faster.  The division is done by the first letter of the 
*untranslated* English name.  However, in some parts of 
objectnamelist.cpp, the list is selected by the first letter of the 
*translated* name.  Whoops.

[[Developer notes]]
Given that the lists are divided by their English name, its pretty 
confusing that the object names are stored in their translated form.  
Why not just store the names in English, since we use English object 
names throughout the code?  This class is in need of some revision for 
KDE4.

BUG: 112925
CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=463659