Fixing bug 113450 (startup time an hour off if daylight savings time active).
This is a symptom of the extremely complex, extremely interconnected, 
extremely fragile startup procedure in kstars-3.5 branch.  The problem 
has been fixed properly in trunk, but I decided at the time that the 
changes were too major to do in the 3.5 branch during a feature freeze.

I have applied a band-aid fix that solves the symptom of this bug; I 
should backport the full startup changes from trunk once 3.5.0 is out.

BUG: 113450
CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=465147