Working in skycomponents; still uncompileable:
+ Refactoring: renamed "AbstractPlanetComponent" to 
"SolarSystemSingleComponent"; this is the base class for components 
which contain one KSPlanetBase object (SunComp, MoonComp, 
PlanetComp, PlutoComp)

+ Refactoring: added "SolarSystemListComponent"; this is the base class 
for components which contain a QList of objects (AsteroidsComp, 
CometsComp)

+ All components draw() function: use KStars* arg instead of SkyMap*, 
because we need access to KStarsData, and there is no SkyMap::data() 

+ Add visible() checks to solar system draw() functions

+ SolarSystemComposite: now contains addComponent()'s for all children.
Thomas had set it up so that the inner bodies (Sun, Moon, Mercury, 
Venus) have separate pointers, in order to draw them in the correct 
order so they appear "behind" each other correctly.  However, I'd rather 
handle this by reimplementing SkyComposite::draw(), if possible.

+ Added draw() functions for EclipticComp, EquatorComp, and PlutoComp



svn path=/trunk/KDE/kdeedu/kstars/; revision=473201