More SkyComposite stuff:
Add ConstellationLinesComposite, which contains a large number of
ConstellationLinesComponents.  Each CLComponent represents a single
series of SkyPoints connected by a continuous line.  Thus, a
constellation figure may be made up of any number of CLComponents.  The
CLComposite ctor reads in the clines.data file, and creates the
CLComponents based on that information.

Added a "bool &found" argument to hasTrail(SkyObject *o, bool &found).
hasTrail loops through all solar system bodies to first find object o,
and then to see if it has a Trail.  It returns true if both of these
conditions are met.  So, a return of false is ambiguous, because you
don't know if it didn't find o, or if o had no Trail.  The "found"
argument gets around this: it is set to true if object o is found.
The practical advantage of this is that the search for o can now be cut
short if it is found and has no Trail, whereas before it would have
had to complete the search in this case.

Add implementations of findByName(QString) to SingleComponent and
ListComponent.  Each object has three names (primary, secondary, and
long); all three names are checked for a match.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=474367