More progress:
EquatorComponent, PointsListComponent, SkyComposite now compile.

SkyMap::checkVisibility() now has only one argument, the SkyPoint to be 
tested.  The other args were redundant, because their values are stored 
internally in SkyMap and updated with SkyMap::setMapGeometry().
I updated all instances of checkVisibility() in skycomponents/, but it 
still needs to be done in the main source directory.

Add public function SkyMap::isSlewing(), which returns true if the sky 
is in motion.  This is now necessary because the SkyComponent::draw() 
functions do not have access to the private SkyMap members "bool 
slewing" and "bool clockSlewing".

Other miscellaneous fixes.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=475061