Adding objectNearest( SkyPoint *p, double &maxrad ) to the Components
This function looks for SkyObjects in the Component that are closer to p 
than maxrad.  If such an object is found, a pointer to it is returned, 
and its angular distance from p is returned-by-reference through maxrad.
If no object is found to be nearer to p than maxrad, then it returns 0.

SkyComposite::objectNearest() executes objectNearest() for each of its 
child Components, and returns the overall nearest SkyObject.

The result is that calling skyComposite()->objectNearest( p, maxrad ) 
will return a pointer to the object that is closest to p.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=480085