Adding replacement for ObjectNamesList.
ObjectNamesList was an array of 27 lists of pointers to SkyObjectName 
objects.  It was designed to provide fast searches of a SkyObject by 
name, and was used by the Find Object tool.  The SkyObjectName class 
stored the name of the object as well as a pointer to the instance of 
the object in the program.

Now that we have the SkyComponents paradigm, this collection of pointers 
is no longer appropriate.  I have replaced the old scheme with a simple 
QStringList of object names, stored in SkyMapComposite.  A SkyComponent 
is responsible for adding its name(s) to the QStringList when it 
initializes itself.  Also, the dtor of the SkyComponent is responsible 
for removing the name(s) from the QStringList.

This is a much simpler scheme than before.  It may not be as fast to 
search for objects in the Find Object tool.  If we find that th new 
implementation is too slow, we can always try a similar optimization to 
what was used by ObjectNamesList (i.e., split the names into 27 lists, 
and select an "active" sub-list as soon as the first letter of the name 
is entered).

TODO: Still need to remove ObjectNameList and SkyObjectName classes, and 
remove vestigal uses of them from the code.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=488367