I believe all files in the main kstars source code directory now compile!
Today I got all skymap* and kstars* files to compile.

On to the tools directory; once that's done we should have an 
executable.

I added a function clearTrailsExcept() to the SkyComponents classes.  
This function removes the trail from all solar system bodies, *except* 
the one specified as an argument.  This is necessary, because when the 
user presses the "Clear Trails" button in the config window, we want to 
keep the "temporary" trail on, if a solar system body is being tracked.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=489302