Well, I've gotten the code to run all the way through KStars::datainitFinished() before crashing.  This is a slot that runs immediately after KStarsData is finished loading all the data from disk.
The problem is, shortly after datainitFinished() has completed, I get a 
crash caused by an index out of bounds in an unnamed QList.  And since 
datainitFinished() is a slot that is triggered at the end of 
KStarsData::slotInitialize(), it isn't easy to figure out exactly where 
the crash is occuring.

My usual trick is to run the program in gdb, because it identifies the 
function in which the crash occurred.  Unfortunately, when running in 
gdb, I don't get a crash at all, rather it seems to enter an infinite 
loop while setting the date and time.  Doesn't happen outside of gdb.  
So that's weird.

I also tried adding a breakpoint in KDevelop, with the hope that I could 
then step through until getting to the crash.  Unfortunately I can't get 
the integrated debugger in KDevelop to work.

So, I'm a bit stuck.  If anyone tries running the current code, please 
post your experiences in this thread.  I guess I'll try inserting some 
debug statements to track down the crash.  Nothing like brute force :/

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=501353