Ran valgrind memcheck to look for leaks.  It only found one real issue, and a few false alarms (I think):
dmsBox:  didn't assign member variable "deg" in setDegType().

KSMoon: false report of mem leak in MoonLRData and MoonBData lists 
(they are cleared correctly in the dtor)

KStarsData::readINDIHosts(): report that xmlParser not deleted, but it
looks like it is (I don't really understand the C-isms in lilxml.c).

KStarsData: false report of mem leak in geoList (it is cleared 
correctly in the dtor)

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=535909