Moved code related to saving config data from KStars::~KStars() to KStars::slotAboutToClose(), a new slot connected to QApplication::aboutToClose().
I don't know why, but it seems that the code in ~KStars() isn't getting 
executed anymore.  I added a kDebug() to the top of the function; it 
never gets reported.  I first noticed the problem because the Startup 
wizard kept opening on every startup...it should open only on the first 
startup, after which it gets suppressed by a config option 
(Options::setRunStartupWizard(false)).  This led me to the realization 
that the Config related code in ~KStars() wasn't getting executed.

CCMAIL: kstars-devel@kde.org
  


svn path=/trunk/KDE/kdeedu/kstars/; revision=536049