Implementing multiple map projections!
You now have a choice in how the sky map is projected:

Lambert Azimuthal Equal-area (this is the one we've been using)
Azimuthal Equidistant (similar to Lambert)
Orthographic (like viewing a globe)
Stereographic
Gnomonic (great circles are straight lines)

You can choose the projection in the View menu, or with the F5-F10 keys
TO get a feel for the differences, try zooming out to see the whole 
projection and then switch between them.

WARNING: I renamed "getXY()" to "toScreen()" and "dXdYToRaDec()" to 
"fromScreen()" to better reflect their complementary neature.  However, 
I haven't yet changed fromScreen() to deal with the multiple map 
projections, so things like mouse events won't work properly unless you 
are using Lambert. 

TODO: the Projection submenu items need to be radio-style items.
 
I also fixed a bug that caused a crash when opening the Configuration 
window.  However, there are still problems with that window (after all, 
this is the first time we've seen it in 4.x!)

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=546242