Last commit (hopefully) before INDI v0.4 is out. Will need to run a few tests.
Please test this updated version to see if it works with your device(s).

I remove Astrophysics mount because there are no testers and it's been
broken for a while.

I noticed that KStars is crawling, I was barely able to do anything,
anyone else experiencing the same?

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=548382