Fix bug causing large lag in Full updates.
The problem was that I was recreating an Earth object inside the loop 
over all copmets and asteroids, even though the Component itself 
contains the Earth object.  Fixing this caused the timing for 
SkyMapComposite::updatePlanets() to go from 1.4 seconds to 0.03 
seconds!!

I also fixed some other minor issues in skycomponents.

Note: I have committed the TIMING code that I had earlier sent to the 
list as patches, so we can continue to use this information for 
profiling.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=549013