Disabling antialiasing.
I had enabled antialiasing for "non-slewing" draw events.  i.e., when 
the sky was in motion, the drawing was not antialiased.  This explains 
why non-slew draws took so much longer than slew draws.

On my system, this change causes a non-slew draw of half the sky to take 
0.3 seconds instead of 1.6.  

It's too bad...I was really looking forward to the antialiased drawing 
in Qt4.  Hopefully, Trolltech will figure out a way to make it faster.  
Perhaps the upcoming QGraphicsView class will help in this regard.  For 
now, I'll leave the antialiasing line commented out.

You'll notice that the star colors look much more saturated after this 
change; this is because the colors get washed out by the antialiasing, 
so I had turned up the saturation to compensate. 

Hopefully, this commit and the previous one will make KStars usable.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=549014