FIxing an unnecessary bottleneck in the draw loop: I had been converting the sky QPixmap to a QImage before calling drawImage() to render it on the SkyMap widget.  However, this conversion takes over 0.1 sec, so now I simply call drawPixmap() instead.
Between this change and making antialiasing optional, KStars should feel 
much "snappier" now.

The time elapsed for the drawing code is still about a factor of two 
longer than it was under 3.5, but we're getting there.  I added detailed 
feedback messages for timing the various draw functions, but they are 
commented out.  If you want to enable them, look for "TIMING" labels in 
the two files modified by this commit, and uncomment the relevant lines.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=550495