Large commit to keep up with kdelibs4_snapshot.  The two big changes are:
1) DCOP has been replaced with DBus.  See my recent post on kstars-devel 
for some pointers on how to prepare your environment for this change.  
You may have to execute kstars with "dbus-launch kstars" if the dbus 
daemon is not already running.

2) KDialog API changes.  For some reason, KStarsSplash was crashing as a 
KDialog, and the backtrace was indicating that something in the size 
calculation was to blame.  I temporarily made KStarsSplash a QDialog 
instead to work around this issue.  I'm reasonably sure it's a problem 
in KDialog, although the rest of our dialogs seem to be working fine...

If you sync your local copy to this update, be aware that you will also 
need the latest qt-copy and kdelibs4_snapshot, and that these now depend 
on dbus-0.62 and QtDBus.  See this post for more info:

http://lists.kde.org/?l=kstars-devel&m=115078593518531&w=2

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=553445